Changing naming styles, formatting
