+++ README.md
@@ -68,7 +68,7 @@ representation of the information, known as plaintext, into an alternative form 
 ##### [Scytale](https://github.com/norkator/Cryptography/blob/master/src/cryptography/ciphers/scytale/Scytale.java)
 ##### [Vigenere](https://github.com/norkator/Cryptography/blob/master/src/cryptography/ciphers/vigenere/Vigenere.java)
 ##### [Elliptic Curve](https://github.com/norkator/Cryptography/blob/master/src/cryptography/ciphers/ellipticCurve/EllipticCurve.java)
-##### [VIC](https://github.com/norkator/Cryptography/blob/master/src/cryptography/ciphers/VIC/VIC.java)
+##### [VIC](https://github.com/norkator/Cryptography/blob/master/src/cryptography/ciphers/vic/VIC.java)
 
 
 Converters
@@ -97,7 +97,7 @@ Encoding is used to represent data in ASCII string format by translating the dat
 ##### [Morse](https://github.com/norkator/Cryptography/blob/master/src/cryptography/encoding/morse/Morse.java)
 ##### [RLE](https://github.com/norkator/Cryptography/blob/master/src/cryptography/encoding/rle/Rle.java)
 ##### [Koblitz](https://github.com/norkator/Cryptography/blob/master/src/cryptography/encoding/koblitz/Koblitz.java)
-##### [VIC sequencing](https://github.com/norkator/Cryptography/blob/master/src/cryptography/encoding/VICsequencing/VICsequencing.java)
+##### [VIC Sequencing](https://github.com/norkator/Cryptography/blob/master/src/cryptography/encoding/vicSequencing/VICSequencing.java)
 
 
 Hashes
